requirements.txt
requests>=2.31.0
supabase==2.27.1
flask>=3.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0